Add tests, move the input parser into its own module
