+++ backend/readme.txt
@@ -1,0 +1,7 @@
+Make a symlink to the engine in this directory, ie
+
+  ln -s ../shadowcraft-engine engine
+
+Run tests with nosetests:
+
+  nosetests -s --rednose tests
